Allow opening difftool on multiple selected files in Commit, Browse, and Diff forms
Except for the staged files in the Commit form, opening the external diff tool
when multiple files were selected would only open it on the first of the
selected files. This change adds support to open the diff tool on all selected
files. This is especially useful when using an external diff tool that supports
opening multiple diffs in a tabbed environment, like Beyond Compare.